docs: fix block preview
