+++ apps/www/.vitepress/theme/components/BlockPage.vue
@@ -2,11 +2,11 @@
 import { useUrlSearchParams } from '@vueuse/core'
 import ComponentLoader from './ComponentLoader.vue'
 
-const params = useUrlSearchParams('hash-params')
+const params = useUrlSearchParams('history')
 </script>
 
 <template>
-  <div v-if="params.name && params.style" :class="params.containerClass">
+  <div v-if="params.name" :class="params.containerClass">
     <ComponentLoader :key="params.style?.toString()" :name="params.name?.toString()" :type-name="'block'" />
   </div>
 </template>
