+++ apps/www/scripts/build-registry.ts
@@ -80,10 +80,17 @@ for (const style of styles) {
       continue
 
     const files = item.files?.map((file) => {
-      let content = fs.readFileSync(
-        path.join(process.cwd(), 'src/lib/registry', style.name, file),
-        'utf8',
-      )
+      let content: string = ''
+
+      try {
+        content = fs.readFileSync(
+          path.join(process.cwd(), 'src/lib/registry', style.name, file),
+          'utf8',
+        )
+      }
+      catch (err) {
+        console.log(`'${file}' is missing`)
+      }
 
       // Replace Windows-style newlines with Unix-style newlines
       content = content.replace(/\r\n/g, newLine)
@@ -99,7 +106,7 @@ for (const style of styles) {
       files,
     }
 
-    const payloadStr = JSON.stringify(payload, null, 2).replace(/\r\n/g, newLine)
+    const payloadStr = `${JSON.stringify(payload, null, 2).replace(/\r\n/g, newLine)}\n`
 
     fs.writeFileSync(
       path.join(targetPath, `${item.name}.json`),
@@ -195,73 +202,73 @@ export const BASE_STYLES = `@tailwind base;
 export const BASE_STYLES_WITH_VARIABLES = `@tailwind base;
 @tailwind components;
 @tailwind utilities;
- 
+
 @layer base {
   :root {
     --background: <%- colors.light["background"] %>;
     --foreground: <%- colors.light["foreground"] %>;
- 
+
     --muted: <%- colors.light["muted"] %>;
     --muted-foreground: <%- colors.light["muted-foreground"] %>;
- 
+
     --popover: <%- colors.light["popover"] %>;
     --popover-foreground: <%- colors.light["popover-foreground"] %>;
- 
+
     --card: <%- colors.light["card"] %>;
     --card-foreground: <%- colors.light["card-foreground"] %>;
- 
+
     --border: <%- colors.light["border"] %>;
     --input: <%- colors.light["input"] %>;
- 
+
     --primary: <%- colors.light["primary"] %>;
     --primary-foreground: <%- colors.light["primary-foreground"] %>;
- 
+
     --secondary: <%- colors.light["secondary"] %>;
     --secondary-foreground: <%- colors.light["secondary-foreground"] %>;
- 
+
     --accent: <%- colors.light["accent"] %>;
     --accent-foreground: <%- colors.light["accent-foreground"] %>;
- 
+
     --destructive: <%- colors.light["destructive"] %>;
     --destructive-foreground: <%- colors.light["destructive-foreground"] %>;
- 
+
     --ring: <%- colors.light["ring"] %>;
- 
+
     --radius: 0.5rem;
   }
- 
+
   .dark {
     --background: <%- colors.dark["background"] %>;
     --foreground: <%- colors.dark["foreground"] %>;
- 
+
     --muted: <%- colors.dark["muted"] %>;
     --muted-foreground: <%- colors.dark["muted-foreground"] %>;
- 
+
     --popover: <%- colors.dark["popover"] %>;
     --popover-foreground: <%- colors.dark["popover-foreground"] %>;
- 
+
     --card: <%- colors.dark["card"] %>;
     --card-foreground: <%- colors.dark["card-foreground"] %>;
- 
+
     --border: <%- colors.dark["border"] %>;
     --input: <%- colors.dark["input"] %>;
- 
+
     --primary: <%- colors.dark["primary"] %>;
     --primary-foreground: <%- colors.dark["primary-foreground"] %>;
- 
+
     --secondary: <%- colors.dark["secondary"] %>;
     --secondary-foreground: <%- colors.dark["secondary-foreground"] %>;
- 
+
     --accent: <%- colors.dark["accent"] %>;
     --accent-foreground: <%- colors.dark["accent-foreground"] %>;
- 
+
     --destructive: <%- colors.dark["destructive"] %>;
     --destructive-foreground: <%- colors.dark["destructive-foreground"] %>;
- 
+
     --ring: <%- colors.dark["ring"] %>;
   }
 }
- 
+
 @layer base {
   * {
     @apply border-border;
@@ -319,64 +326,64 @@ export const THEME_STYLES_WITH_VARIABLES = `
   .theme-<%- theme %> {
     --background: <%- colors.light["background"] %>;
     --foreground: <%- colors.light["foreground"] %>;
- 
+
     --muted: <%- colors.light["muted"] %>;
     --muted-foreground: <%- colors.light["muted-foreground"] %>;
- 
+
     --popover: <%- colors.light["popover"] %>;
     --popover-foreground: <%- colors.light["popover-foreground"] %>;
- 
+
     --card: <%- colors.light["card"] %>;
     --card-foreground: <%- colors.light["card-foreground"] %>;
- 
+
     --border: <%- colors.light["border"] %>;
     --input: <%- colors.light["input"] %>;
- 
+
     --primary: <%- colors.light["primary"] %>;
     --primary-foreground: <%- colors.light["primary-foreground"] %>;
- 
+
     --secondary: <%- colors.light["secondary"] %>;
     --secondary-foreground: <%- colors.light["secondary-foreground"] %>;
- 
+
     --accent: <%- colors.light["accent"] %>;
     --accent-foreground: <%- colors.light["accent-foreground"] %>;
- 
+
     --destructive: <%- colors.light["destructive"] %>;
     --destructive-foreground: <%- colors.light["destructive-foreground"] %>;
- 
+
     --ring: <%- colors.light["ring"] %>;
- 
+
     --radius: 0.5rem;
   }
- 
+
   .dark .theme-<%- theme %> {
     --background: <%- colors.dark["background"] %>;
     --foreground: <%- colors.dark["foreground"] %>;
- 
+
     --muted: <%- colors.dark["muted"] %>;
     --muted-foreground: <%- colors.dark["muted-foreground"] %>;
- 
+
     --popover: <%- colors.dark["popover"] %>;
     --popover-foreground: <%- colors.dark["popover-foreground"] %>;
- 
+
     --card: <%- colors.dark["card"] %>;
     --card-foreground: <%- colors.dark["card-foreground"] %>;
- 
+
     --border: <%- colors.dark["border"] %>;
     --input: <%- colors.dark["input"] %>;
- 
+
     --primary: <%- colors.dark["primary"] %>;
     --primary-foreground: <%- colors.dark["primary-foreground"] %>;
- 
+
     --secondary: <%- colors.dark["secondary"] %>;
     --secondary-foreground: <%- colors.dark["secondary-foreground"] %>;
- 
+
     --accent: <%- colors.dark["accent"] %>;
     --accent-foreground: <%- colors.dark["accent-foreground"] %>;
- 
+
     --destructive: <%- colors.dark["destructive"] %>;
     --destructive-foreground: <%- colors.dark["destructive-foreground"] %>;
- 
+
     --ring: <%- colors.dark["ring"] %>;
   }`
